requirements.txt
APScheduler==3.6.3
attrs==20.2.0
Automat==20.2.0
bitly-api==0.3
boto3==1.15.16
botocore==1.18.16
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
decorator==4.4.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
gspread==3.6.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.1
itemloaders==1.0.3
itsdangerous==1.1.0
jmespath==0.10.0
lxml==4.5.2
MarkupSafe==1.1.1
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
parsel==1.6.0
Pillow==7.2.0
Protego==0.1.16
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
queuelib==1.5.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
schedule==0.6.0
Scrapy==2.3.0
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.20
tornado==6.0.4
Twisted==20.3.0
tzlocal==2.1
urllib3==1.25.10
w3lib==1.22.0
Werkzeug==1.0.1
zope.interface==5.1.2
